requirements.txt
matplotlib==3.10.6
numpy==2.3.2
pillow==11.3.0
requests==2.32.5
python-dotenv==1.1.1
huggingface_hub==0.34.4
transformers==4.56.0
evaluate==0.4.5
torch==2.8.0
opencv-python==4.12.0.88
tqdm==4.67.1
urllib3==2.3.0